backend/pyproject.toml
[project]
name = "daytistics-backend"
version = "0.1.0"
description = "Backend for Daytistics"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioinject>=0.35.3",
    "aiopg>=1.4.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "asgi-csrf>=0.11",
    "asyncpg>=0.30.0",
    "fastapi-mail>=1.4.2",
    "fastapi[standard]>=0.115.5",
    "freezegun>=1.5.1",
    "greenlet>=3.1.1",
    "jinja2>=3.1.4",
    "passlib[bcrypt]>=1.7.4",
    "polyfactory>=2.18.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.7.0",
    "pyjwt[crypto]>=2.10.1",
    "pytest-sqlalchemy-mock>=0.1.7",
    "strawberry-graphql[fastapi]>=0.253.0",
    "strawberry-sqlalchemy-mapper>=0.5.0",
    "uvicorn>=0.32.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.8.1",
    "faker>=33.1.0",
    "pytest-env>=1.1.5",
    "pytest-asyncio>=0.25.0",
    "asgi-lifespan>=2.1.0",
]
